+++ Main.sublime-menu
@@ -13,7 +13,7 @@
                     { "caption": "Diff Single Override...", "command": "override_audit_diff_single"},
                     { "caption": "Bulk Diff: Single Package...", "command": "override_audit_diff_single", "args": {"bulk": true}},
                     { "caption": "Bulk Diff: All Packages", "command": "override_audit_diff_report"},
-                    { "caption": "Refresh Current Report", "command": "override_audit_context_report"},
+                    { "caption": "Refresh Current Report", "command": "override_audit_refresh_report"},
                     { "caption": "-" }
                 ]
             }
