+++ src/commands/__init__.py
@@ -1,8 +1,8 @@
 from ...bootstrap import reload
 
 reload("src.commands", ["package_report", "override_report", "diff_report",
-       "diff_single", "toggle_override", "diff_override", "edit_override",
-       "delete_override", "freshen_override", "diff_package",
+       "refresh_report", "diff_single", "toggle_override", "diff_override",
+       "edit_override", "delete_override", "freshen_override", "diff_package",
        "freshen_package", "modify_mark"])
 
 from .package_report import OverrideAuditPackageReportCommand
@@ -16,6 +16,7 @@ from .delete_override import OverrideAuditDeleteOverrideCommand
 from .freshen_override import OverrideAuditFreshenOverrideCommand
 from .diff_package import OverrideAuditDiffPackageCommand
 from .freshen_package import OverrideAuditFreshenPackageCommand
+from .refresh_report import OverrideAuditRefreshReportCommand
 from .modify_mark import OverrideAuditModifyMarkCommand
 
 __all__ = [
@@ -23,6 +24,7 @@ __all__ = [
     "OverrideAuditPackageReportCommand",
     "OverrideAuditOverrideReportCommand",
     "OverrideAuditDiffReportCommand",
+    "OverrideAuditRefreshReportCommand",
 
     # Override commands
     "OverrideAuditToggleOverrideCommand",
